+++ src/main/java/frc/robot/RobotContainer.java
@@ -288,7 +288,7 @@ public class RobotContainer {
     IS_L2.and(m_driverController.leftTrigger()).onTrue(prepDealgaeLow());
     IS_L3.and(m_driverController.leftTrigger()).onTrue(prepDealgaeHigh());
 
-    m_driverController.leftTrigger().onFalse(stow());
+    m_driverController.leftTrigger().onFalse(dealgae());
 
     IS_L4
         .and(m_driverController.rightTrigger())
@@ -477,21 +477,25 @@ public class RobotContainer {
   public Command prepDealgaeLow() {
     return Commands.parallel(
         elevator.setHeight(ElevatorConstants.DEALGAE_LOW_POSITION),
-        arm.setAngle(ArmConstants.DEALGAE_LOW_ANGLE),
+        arm.setAngle(ArmConstants.PRE_DEALGAE_ANGLE),
         endEffector.setVoltage(6));
   }
 
   public Command prepDealgaeHigh() {
     return Commands.parallel(
         elevator.setHeight(ElevatorConstants.DEALGAE_HIGH_POSITION),
-        arm.setAngle(ArmConstants.DEALGAE_HIGH_ANGLE),
+        arm.setAngle(ArmConstants.PRE_DEALGAE_ANGLE),
         endEffector.setVoltage(6));
   }
 
-  public Command stow() {
-    return Commands.parallel(
-        elevator.setHeight(ElevatorConstants.STOW_METER),
-        arm.setAngle(ArmConstants.ARM_STOW_ANGLE),
-        endEffector.stopMotor());
+  public Command dealgae() {
+    return Commands.sequence(
+            arm.setAngle(ArmConstants.DEALGAE_ANGLE).withTimeout(0.2),
+            drivetrain.driveBackward(1).withTimeout(0.6),
+            Commands.parallel(
+                elevator.setHeight(ElevatorConstants.STOW_METER),
+                arm.setAngle(ArmConstants.ARM_STOW_ANGLE),
+                endEffector.stopMotor()))
+        .withName("Dealgae");
   }
 }
