+++ src/main/java/frc/robot/subsystems/arm/ArmConstants.java
@@ -20,6 +20,6 @@ public class ArmConstants {
   public static final double ARM_INTAKE_ANGLE = 105.0;
   public static final double ARM_STOW_ANGLE = 90.0;
 
-  public static final double DEALGAE_LOW_ANGLE = 330;
-  public static final double DEALGAE_HIGH_ANGLE = 310;
+  public static final double PRE_DEALGAE_ANGLE = 335;
+  public static final double DEALGAE_ANGLE = 340;
 }
