+++ src/main/java/frc/robot/subsystems/elevator/ElevatorConstants.java
@@ -32,7 +32,7 @@ public class ElevatorConstants {
   public static final double INTAKE_METER = Units.inchesToMeters(9.825);
   public static final double DEALGAE_LOW_POSITION = 0;
   public static final double DEALGAE_HIGH_POSITION =
-      ElevatorSubsystem.rotationsToMeters(17 / GEAR_RATIO_DIFF);
+      ElevatorSubsystem.rotationsToMeters(30 / GEAR_RATIO_DIFF);
   public static final double DEALGAE_STOW_POSITION = 0;
 
   public static final double STOW_METER = ElevatorSubsystem.rotationsToMeters(15 / GEAR_RATIO_DIFF);
